Refactor Image Directories
Move image files into new directories according to image type.
